things done almost

+++ Server/models/nic.js
@@ -2,7 +2,7 @@
 module.exports = (sequelize, DataTypes) => {
   var NICs = sequelize.define('NICs', {
     name: DataTypes.STRING,
-    nicNo: DataTypes.INTEGER
+    nicNo: DataTypes.DOUBLE
   }, {});
   NICs.associate = function(models) {
     // associations can be defined here
